05 commit: fixes and calculating target heart rate in HeartRates.cpp

+++ Ex 3.16 Target-Heart-Rate Calculator/HeartRates.h
@@ -36,7 +36,7 @@ public:
     string getDateOfBirth();
     int getAge();
     int getMaximumHeartRate();
-    int getTargetHeartRate();
+    float getTargetHeartRate();
     
     virtual ~HeartRates();
 private:
